Learn some C preprocessor magic
Despite C preprocessor is a very limited language, it turns out it's
possible to make employ it for semiautomatic unit test registry generation.

In [1] it's explained how to implement 'map' function for processing
variable length macro arguments. Use that implementation, but protect with
the library's custom prefix.

Moreover, clear reference to [1] must be added to the library documentation
in (near) future.

[1]: http://jhnet.co.uk/articles/cpp_magic

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>