Refactor everything a bit
This commit was done after long time passed since the changes were made,
so it's hard to remember what exactly was done. However, the project
architecture is still in its early stage, so that's not a disaster to
have such stupid commits. At least, the project can be successfully built.

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>